Cannot have python floats intermixing with tensor values
Here's the warning:
```
retinaface/layers/functions/prior_box.py:17: TracerWarning: Converting a tensor to a Python float might cause the trace to be incorrect. We can't record the data flow of Python values, so this value will be treated as a constant in the future. This means that the trace might not generalize to other inputs!
math.ceil(self.image_size[0]/step), math.ceil(self.image_size[1]/step)] for step in self.steps
```