Torch tensors are registered as constants...
But some parameters are already tensors. Here's the warning:

```
retinaface/layers/functions/prior_box.py:17: TracerWarning: torch.tensor results are registered as constants in the trace. You can safely ignore this warning if you use this function to create tensors out of constant variables that would be the same every time you call this function. In any other case, this might cause the trace to be incorrect.
  ) / torch.tensor(self.steps).reshape(-1, 1)
```